Add files via upload
The program so far. I need to add validation and finish up the ramp and arch methods/classes. The ramp class is almost done, but I still need to begin the arch method. Adding cost/validation/ and joption will be fairly simple. Figuring out these calculations is proving to be difficult.